requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
anywidget==0.9.13
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
bqplot==0.12.44
branca==0.8.1
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colour==0.1.5
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
duckdb==1.1.3
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.17.0
folium==0.19.4
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2025.2.0
gdown==5.2.0
geographiclib==2.0
geojson==3.2.0
geopandas==1.0.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.11.3
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.19.2
ipython==8.18.1
ipytree==0.2.2
ipyvue==1.11.2
ipyvuetify==1.10.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-leaflet==0.19.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.3
leafmap==0.42.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.24.1
networkx==3.2.1
numpy==1.26.4
openai==1.60.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
plotly==6.0.0
postgrest==0.19.3
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
psygnal==0.11.1
pure_eval==0.2.3
pyarrow==19.0.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.6.1
pyshp==2.3.1
PySocks==1.7.1
pystac==1.10.1
pystac-client==0.8.3
python-box==7.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
realtime==2.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scooby==0.10.0
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
storage3==0.11.3
streamlit==1.41.1
streamlit_folium==0.24.0
StrEnum==0.4.15
supabase==2.12.0
supafunc==0.9.3
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==14.2
whitebox==2.3.5
whiteboxgui==2.3.0
widgetsnbextension==4.0.13
xyzservices==2025.1.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
